fix(sm12x): int64 block offsets in indexer paged-MQA-logits kernel
The post-rebase indexer kv_cache block stride grew (~1039680, a strided slice
of the fused KV block); block_idx * stride overflowed int32 in the SM120 paged
MQA-logits Triton kernel for block_idx beyond ~2065, giving an illegal memory
access on SM120 and silent garbage on SM121 under long-context / multi-request
indexer calls. Cast block_idx to int64 at the K and scale gather sites in
_fp8_paged_mqa_logits_rowwise_kernel.

+++ vllm/models/deepseek_v4/nvidia/ops/sm12x_mqa.py
@@ -373,7 +373,7 @@ def _fp8_paged_mqa_logits_rowwise_kernel(
     )
 
     scale = tl.load(
-        scale_ptr + block_idx * stride_sb + block_offset * stride_ss,
+        scale_ptr + block_idx.to(tl.int64) * stride_sb + block_offset * stride_ss,
         mask=context_mask,
         other=0.0,
     )
@@ -396,7 +396,7 @@ def _fp8_paged_mqa_logits_rowwise_kernel(
             ).to(tl.float32)
             k = tl.load(
                 kv_ptr
-                + block_idx[None, :] * stride_kvb
+                + block_idx[None, :].to(tl.int64) * stride_kvb
                 + block_offset[None, :] * stride_kvs
                 + d[:, None] * stride_kvd,
                 mask=context_mask[None, :] & (d[:, None] < head_dim),
